refactor(#2486): split pressure_drop_interface.py (1424 LOC) into domain modules
Decomposes the monolithic pressure_drop_interface.py into three focused sub-modules:
- _pressure_drop_helpers.py (429 LOC): listing/info helpers + _wrap_text utility
- _pressure_drop_validation.py (230 LOC): input validation functions
- _pressure_drop_output.py (320 LOC): output formatting functions
- pressure_drop_interface.py (515 LOC coordinator): core calculation functions +
  main() CLI + re-exports for backward compat

Cross-module _wrap_text text-wrapping utility moved to _pressure_drop_helpers.py,
imported by both validation and output. main() CLI kept in coordinator to avoid
circular import with calculate_pressure_drop*.

TDD: 7/7 structure/LOC tests pass; 5 API tests skipped (sympy not installed).
